imp: organize ack types under the ics04
