+++ .changelog/unreleased/breaking-changes/717-organize-ack-types-under-ics04.md
@@ -1,0 +1,3 @@
+- Gather all the acknowledgement types and organize them under
+  the `ics04_channel` module to get accessible by any applications.
+  ([#717](https://github.com/cosmos/ibc-rs/issues/717))
