+++ crates/ibc/src/core/ics04_channel/mod.rs
@@ -11,6 +11,7 @@ pub mod msgs;
 pub mod packet;
 pub mod timeout;
 
+pub mod acknowledgement;
 pub mod commitment;
 mod version;
 pub use version::Version;
